Moved tests into runtime and fixed up
I've moved the tests and the conftest file into runtime. Now tests pass as well as deploys. It was not easy, and runtime tests now have to be run from a separate directory - coming up with unified pytest regime when the CDK purposefully ignores all parent directories is beyond me at the moment. However, it does work. And that's probably a good thing. Testing in CI/CD may yet prove to be my undoing
